requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
aiosqlite
sqlmodel
nats-py
playwright
aiogram
python-dotenv